pyproject.toml
[tool.poetry]
name = "pgmigrate"
version = "0.1.0"
description = ""
authors = ["Shivang Rathore <rshivang12345@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
asyncpg = "^0.29.0"
click = "^8.1.7"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
toml = "^0.10.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
